Move PageHandler to layout component
