+++ src/layouts/ViewRouter/ViewRouter.js
@@ -9,7 +9,7 @@ import EventDetailView from '../../views/EventDetailView';
 import AddressView from '../../views/AddressView';
 import ServiceTreeView from '../../views/ServiceTreeView';
 import ViewTitle from './ViewTitle';
-import PageHandler from '../../views/components/PageHandler';
+import PageHandler from '../PageHandler';
 import FeedbackView from '../../views/FeedbackView';
 import DivisionView from '../../views/DivisionView';
 import InfoView from '../../views/InfoView';
